Update the output sub.sample (hidden)
The output collection of sub.sample is hidden . This question "What would you expect the result of count.groups on this new shared output collection to be? Check if you are correct." is unclear which input collection should be using for running the "count.groups" . The input collection needs to unhide (named 0.03) and use as the input to run count.groups.